+++ src/server.hpp
@@ -184,6 +184,14 @@ void server_on_submit_job(ServerData * data,
                           IPMessage * task_data);
 
 /**
+ * @brief Server PROFILE_SUBMITTED_BY_DP handler
+ * @param[in,out] data The data associated with the server_process
+ * @param[in,out] task_data The data associated with the message the server received
+ */
+void server_on_submit_profile(ServerData * data,
+                               IPMessage * task_data);
+
+/**
  * @brief Server SCHED_REJECT_JOB handler
  * @param[in,out] data The data associated with the server_process
  * @param[in,out] task_data The data associated with the message the server received