+++ tools/check_batsim_version.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 """Check that batsim --version is consistent with batsim git releases."""
 
+import argparse
 import re
 import subprocess
 import sys
@@ -22,14 +23,27 @@ def version_from_string(input_string,
 def batsim_binary_version(batsim_command="batsim",
                           pattern='''(v\d+\.\d+\.\d+).*'''):
     """Retrieve the version from the batsim binary."""
-    p = subprocess.run("{cmd} --version".format(cmd=batsim_command),
-                       shell=True, check=True, stdout=subprocess.PIPE)
+    cmd = "{cmd} --version".format(cmd=batsim_command)
+    p = subprocess.run(cmd, shell=True, check=True,
+                       stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
-    return version_from_string(p.stdout.decode('utf-8'), pattern)
+    if p.returncode != 0:
+        raise Exception("Command '{cmd}' returned {ret}.\n"
+                        "Output:\n{o}\n\n"
+                        "Error:\n{e}\n".format(cmd=cmd,
+                                               ret=p.returncode,
+                                               o=p.stdout.decode('utf-8'),
+                                               e=p.stderr.decode('utf-8')))
+
+    binary_version = p.stdout.decode('utf-8').strip()
+    print("Batsim binary full version: {}".format(binary_version))
+    return version_from_string(binary_version, pattern)
 
 
 def batsim_git_version(
-        git_command='git describe --match "v[0-9]*.[0-9]*.[0-9]*"',
+        git_command="git",
+        git_path=".",
+        git_subcommand='describe --match "v[0-9]*.[0-9]*.[0-9]*"',
         pattern='''(v\d+\.\d+\.\d+).*'''):
     """Retrieve the latest released Batsim version (in the git tags).
 
@@ -39,20 +53,41 @@ def batsim_git_version(
     - v1.1.0-20-g0fcd0c9  if the current commit is after v1.1.0 in the tree
     - or fail if there is no released version before the current commit.
     """
-    p = subprocess.run("{cmd}".format(cmd=git_command),
-                       shell=True, stdout=subprocess.PIPE)
+    cmd = "{cmd} -C {p} {scmd}".format(cmd=git_command,
+                                       p=git_path,
+                                       scmd=git_subcommand)
+    p = subprocess.run(cmd, shell=True, check=True,
+                       stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
     if p.returncode != 0:
-        raise Exception("Command '{cmd}' returned {ret}. Output:\n{o}".format(
-            cmd=git_command, ret=p.returncode, o=p.stdout.decode('utf-8')))
+        raise Exception("Command '{cmd}' returned {ret}.\n"
+                        "Output:\n{o}\n\n"
+                        "Error:\n{e}\n".format(cmd=cmd,
+                                               ret=p.returncode,
+                                               o=p.stdout.decode('utf-8'),
+                                               e=p.stderr.decode('utf-8')))
 
-    return version_from_string(p.stdout.decode('utf-8'), pattern)
+    git_version = p.stdout.decode('utf-8').strip()
+    print("git full version: {}".format(git_version))
+    return version_from_string(git_version, pattern)
 
 
 def main():
     """Entry point."""
-    bin_version = batsim_binary_version()
-    git_version = batsim_git_version()
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument('--batsim-command',
+                        type=str,
+                        default='batsim',
+                        help='The batsim command (without --version)')
+    parser.add_argument('--batsim-git-dir',
+                        type=str,
+                        default='.',
+                        help='The batsim git directory')
+
+    args = parser.parse_args()
+
+    bin_version = batsim_binary_version(batsim_command=args.batsim_command)
+    git_version = batsim_git_version(git_path=args.batsim_git_dir)
 
     if bin_version == git_version:
         print("Batsim binary and git match on version {v}".format(